+++ serviceTest/src/test/scala/org/broadinstitute/dsde/workbench/service/Rawls.scala
@@ -66,6 +66,11 @@ trait Rawls extends RestClient with LazyLogging {
       mapper.readTree(response).at("/workspace/bucketName").asText()
     }
 
+    def getWorkflowCollectionName(namespace: String, name: String)(implicit token: AuthToken): String = {
+      val response = parseResponse(getRequest(url + s"api/workspaces/$namespace/$name"))
+      mapper.readTree(response).at("/workspace/workflowCollectionName").asText()
+    }
+
     def list()(implicit token: AuthToken):String  = {
       logger.info(s"Listing workspaces")
       parseResponse(getRequest(url + s"api/workspaces"))